Check fasta index separately than fasta file

+++ nextflow/vcf_prepper/bin/process_fasta.py
@@ -57,10 +57,17 @@ def bgzip_fasta(fasta_dir: str, unzipped_fasta: str) -> str:
 
     return unzipped_fasta + ".gz"
 
-def index_fasta(zipped_fasta: str) -> None:
+def index_fasta(zipped_fasta: str, force: str = False) -> None:
     if not os.path.isfile(zipped_fasta):
         print(f"[ERROR] Cannot index fasta - {fasta} - does not exist. Exiting ...")
         exit(1)
+
+    fai = os.path.join(zipped_fasta, ".fai")
+    gzi = os.path.join(zipped_fasta, ".gzi")
+
+    if os.path.isfile(fai) and os.path.isfile(gzi) and not force:
+        print(f"[INFO] both .fai and .gzi file exist. Skipping ...")
+        return
 
     if os.path.isfile(zipped_fasta + ".fai"):
         print(f"[INFO] {zipped_fasta + '.fai'} exist. Deleting ...")
@@ -99,41 +106,48 @@ def main(args = None):
     
     fasta_dir = args.fasta_dir or FASTA_DIR
     fasta_glob = os.path.join(fasta_dir, f"{fasta_species_name}.{assembly}.dna.*.fa.gz")
-    if glob.glob(fasta_glob):
-        if not args.force:
-            print(f"[INFO] {fasta_glob} exists. Skipping ...")
-            exit(0)
-        else:
-            # for human we check and delete fasta manually if needed 
-            if species.startswith("homo_sapiens"):
-                print(f"[ERROR] {fasta_glob} exists for human. Won't be overwritten ...")
-                exit(1)
+    
+    if len(glob.glob(fasta_glob)) > 1:
+        print(f"[ERROR] multiple fasta file found. Check {fasta_glob}")
+        exit(1)
 
+    fasta = None
+    if glob.glob(fasta_glob) and not args.force:
+        print(f"[INFO] {fasta_glob} exists. Skipping ...")
+        fasta = glob.glob(fasta_glob)[0]
+    elif glob.glob(fasta_glob) and species.startswith("homo_sapiens"):
+        # for human we check and delete fasta manually if needed 
+        print(f"[WARN] {fasta_glob} exists for human. Won't be overwritten ...")
+        fasta = glob.glob(fasta_glob)[0]
+    else:
+        if glob.glob(fasta_glob):
             print(f"[INFO] {fasta_glob} exists. Will be oerwritten ...")
             for f in glob.glob(fasta_glob):
                 os.remove(f)
         
-    rl_version = get_relative_version(version, division)
-    src_compressed_fasta = get_ftp_path(species, assembly, division, rl_version, "fasta", "local", fasta_species_name)
+        rl_version = get_relative_version(version, division)
+        src_compressed_fasta = get_ftp_path(species, assembly, division, rl_version, "fasta", "local", fasta_species_name)
     
-    if src_compressed_fasta is not None:
-        compressed_fasta = os.path.join(fasta_dir, os.path.basename(src_compressed_fasta))
-        returncode = copyto(src_compressed_fasta, compressed_fasta)
+        if src_compressed_fasta is not None:
+            compressed_fasta = os.path.join(fasta_dir, os.path.basename(src_compressed_fasta))
+            returncode = copyto(src_compressed_fasta, compressed_fasta)
     
-    if src_compressed_fasta is None or returncode != 0:
-        print(f"[INFO] Failed to copy fasta file - {src_compressed_fasta}, will retry using remote FTP")
+        if src_compressed_fasta is None or returncode != 0:
+            print(f"[INFO] Failed to copy fasta file - {src_compressed_fasta}, will retry using remote FTP")
         
-        compressed_fasta_url = get_ftp_path(species, assembly, division, rl_version, "fasta", "remote", fasta_species_name)
+            compressed_fasta_url = get_ftp_path(species, assembly, division, rl_version, "fasta", "remote", fasta_species_name)
         
-        compressed_fasta = os.path.join(fasta_dir, compressed_fasta_url.split('/')[-1])
-        returncode = download_file(compressed_fasta, compressed_fasta_url)
-        if returncode != 0:
-            print(f"[ERROR] Could not download fasta file - {compressed_fasta_url}")
-            exit(1)
+            compressed_fasta = os.path.join(fasta_dir, compressed_fasta_url.split('/')[-1])
+            returncode = download_file(compressed_fasta, compressed_fasta_url)
+            if returncode != 0:
+                print(f"[ERROR] Could not download fasta file - {compressed_fasta_url}")
+                exit(1)
     
-    unzipped_fasta = ungzip_fasta(fasta_dir, compressed_fasta)
-    zipped_fasta = bgzip_fasta(fasta_dir, unzipped_fasta)
-    index_fasta(zipped_fasta)
+        unzipped_fasta = ungzip_fasta(fasta_dir, compressed_fasta)
+        fasta = bgzip_fasta(fasta_dir, unzipped_fasta)
+    
+    if fasta is not None:
+        index_fasta(fasta, force=args.force)
         
 if __name__ == "__main__":
     sys.exit(main())
